Live validation: race conditions in async validators on same node is now a non-issue
Removed the concept of an error handler, and now we have a change map instead.
It makes it much easier to create a UI that responds to fields that goes
from validated -> not-validated -> validated

All changes in the dependency graph from a single point
is tracked, and returned in the change map.